reafactor: get navbar user into utils folder
