+++ src/pages/dashboard.js
@@ -1,30 +1,9 @@
-import NavbarAtlet from "@/components/Navbar/NavbarAtlet";
-import NavbarPelatih from "@/components/Navbar/NavbarPelatih";
-import NavbarUmpire from "@/components/Navbar/NavbarUmpire";
-import React, { useState } from "react";
+import React from "react";
 
-const Dashboard = () => {
-	const [user, setUser] = useState("ATLET");
-
-	const getNavbarUser = (user) => {
-		switch (user) {
-			case "ATLET":
-				return <NavbarAtlet />;
-			case "PELATIH":
-				return <NavbarPelatih />;
-			case "UMPIRE":
-				return <NavbarUmpire />;
-			default:
-				return <h1>restricted</h1>;
-		}
-	};
-
+export default function Dashboard() {
 	return (
 		<>
-			{getNavbarUser("ATLET")}
-			<></>
+			<h1>ini page dashboard</h1>
 		</>
 	);
-};
-
-export default Dashboard;
+}
